Added save logic, united ribbons to 1 common ribbon (#6)
* Added save logic, united ribbons to 1 common ribbon

* Updated assets dir name+ updated workload icon

* Refactor import order in HelloWorldItemEditor for consistency
