+++ Workload/app/items/HelloWorldItem/HelloWorldItemEditor.tsx
@@ -1,20 +1,19 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useLocation } from "react-router-dom";
+import { Stack } from "@fluentui/react";
 import { useTranslation } from "react-i18next";
 import { PageProps, ContextProps } from "../../App";
-import { ItemWithDefinition, getWorkloadItem } from "../../controller/ItemCRUDController";
+import { ItemWithDefinition, getWorkloadItem, callGetItem, saveItemDefinition } from "../../controller/ItemCRUDController";
+import { callOpenSettings } from "../../controller/SettingsController";
+import { callNotificationOpen } from "../../controller/NotificationController";
 import { ItemEditorLoadingProgressBar } from "../../controls/ItemEditorLoadingProgressBar";
-import { HelloWorldItemDefinition } from "./HelloWorldItemModel";
+import { HelloWorldItemDefinition, VIEW_TYPES, CurrentView } from "./HelloWorldItemModel";
 import { HelloWorldItemEditorEmpty } from "./HelloWorldItemEditorEmpty";
-import { HelloWorldItemEditorGettingStarted } from "./HelloWorldItemEditorGettingStarted";
+import { HelloWorldItemEditorDefault } from "./HelloWorldItemEditorDefault";
 import "../../styles.scss";
+import { HelloWorldItemRibbon } from "./HelloWorldItemRibbon";
 
-/**
- * Main editor component for HelloWorld item.
- *
- * Note: To remove the empty state, simply set defaultView to 'getting-started'
- * or modify the logic in line 62-73
- */
+
 export function HelloWorldItemEditor(props: PageProps) {
   const { workloadClient } = props;
   const pageContext = useParams<ContextProps>();
@@ -23,80 +22,109 @@ export function HelloWorldItemEditor(props: PageProps) {
   // State management
   const [isLoading, setIsLoading] = useState(true);
   const [item, setItem] = useState<ItemWithDefinition<HelloWorldItemDefinition>>();
-  const [currentView, setCurrentView] = useState<'empty' | 'getting-started'>('empty');
+  const [currentView, setCurrentView] = useState<CurrentView>(VIEW_TYPES.EMPTY);
+  const [hasBeenSaved, setHasBeenSaved] = useState<boolean>(false);
 
-  // Get storage key for this specific item
-  const getStorageKey = (itemId: string) => `helloworld-editor-state-${itemId}`;
   const { pathname } = useLocation();
 
   async function loadDataFromUrl(pageContext: ContextProps, pathname: string): Promise<void> {
-      setIsLoading(true);
-      var item: ItemWithDefinition<HelloWorldItemDefinition> = undefined;
-      if (pageContext.itemObjectId) {
-        // for Edit scenario we get the itemObjectId and then load the item via the workloadClient SDK
-        try {
-          item = await getWorkloadItem<HelloWorldItemDefinition>(
-            workloadClient,
-            pageContext.itemObjectId,
-          );
+    setIsLoading(true);
+    var LoadedItem: ItemWithDefinition<HelloWorldItemDefinition> = undefined;
+    if (pageContext.itemObjectId) {
+      // for Edit scenario we get the itemObjectId and then load the item via the workloadClient SDK
+      try {
+        LoadedItem = await getWorkloadItem<HelloWorldItemDefinition>(
+          workloadClient,
+          pageContext.itemObjectId,
+        );
 
-          // Ensure item definition is properly initialized without mutation
-          if (!item.definition) {
-            item = {
-              ...item,
-              definition: {
-                message: undefined,
-              }
-            };
-          }
-          setItem(item);
-        } catch (error) {
-          setItem(undefined);
+        // Ensure item definition is properly initialized without mutation
+        if (!LoadedItem.definition) {
+          LoadedItem = {
+            ...LoadedItem,
+            definition: {
+              state: undefined,
+            }
+          };
         }
-      } else {
-        console.log(`non-editor context. Current Path: ${pathname}`);
-      }
-      setIsLoading(false);
-    }
+        else {
+          console.log('LoadedItem definition: ', LoadedItem.definition);
+        }
 
-  useEffect(() => {
-    if (pageContext?.itemObjectId) {
-      const stored = sessionStorage.getItem(getStorageKey(pageContext.itemObjectId));
-      if (stored === 'getting-started' || stored === 'empty') {
-        setCurrentView(stored);
-      } else {
-        setCurrentView('empty'); // default for new items
+        setItem(LoadedItem);
+        setCurrentView(!LoadedItem?.definition?.state ? VIEW_TYPES.EMPTY : VIEW_TYPES.GETTING_STARTED);
+
+      } catch (error) {
+        setItem(undefined);
       }
     } else {
-      setCurrentView('empty');
+      console.log(`non-editor context. Current Path: ${pathname}`);
     }
-  }, [pageContext?.itemObjectId]);
+    setIsLoading(false);
+  }
 
   useEffect(() => {
-        loadDataFromUrl(pageContext, pathname);
-      }, [pageContext, pathname]);
+    setHasBeenSaved(false);
+  }, [currentView, item?.id]);
 
-  /**
-   * Save current view state to sessionStorage
-   */
-  const saveViewState = (view: 'empty' | 'getting-started') => {
-    if (pageContext?.itemObjectId) {
-      sessionStorage.setItem(getStorageKey(pageContext.itemObjectId), view);
+  useEffect(() => {
+    loadDataFromUrl(pageContext, pathname);
+  }, [pageContext, pathname]);
+
+
+  const navigateToGettingStarted = () => {
+    setCurrentView(VIEW_TYPES.GETTING_STARTED);
+  };
+
+  const handleOpenSettings = async () => {
+    if (item) {
+      try {
+        const item_res = await callGetItem(workloadClient, item.id);
+        await callOpenSettings(workloadClient, item_res.item, 'About');
+      } catch (error) {
+        console.error('Failed to open settings:', error);
+      }
     }
   };
 
-  const navigateToGettingStarted = () => {
-    setCurrentView('getting-started');
-    saveViewState('getting-started');
+  async function SaveItem() {
+    var successResult = await saveItemDefinition<HelloWorldItemDefinition>(
+      workloadClient,
+      item.id,
+      {
+        state: VIEW_TYPES.GETTING_STARTED
+      });
+    const wasSaved = Boolean(successResult);
+    setHasBeenSaved(wasSaved);
+    callNotificationOpen(
+      props.workloadClient,
+      t("ItemEditor_Saved_Notification_Title"),
+      t("ItemEditor_Saved_Notification_Text", { itemName: item.displayName }),
+      undefined,
+      undefined
+    );
+  }
+
+  const isSaveEnabled = () => {
+    if (currentView === VIEW_TYPES.EMPTY) {
+      return false;
+    }
+
+    if (currentView === VIEW_TYPES.GETTING_STARTED) {
+      if (hasBeenSaved) {
+        return false;
+      }
+
+      if (!item?.definition?.state) {
+        return true;
+      }
+
+      return false;
+    }
+
+    return false;
   };
 
-  /**
-   * Navigate back to Empty view
-   */
-  const navigateToEmpty = () => {
-    setCurrentView('empty');
-    saveViewState('empty');
-  };
 
   // Show loading state
   if (isLoading) {
@@ -109,20 +137,27 @@ export function HelloWorldItemEditor(props: PageProps) {
 
   // Render appropriate view based on state
   return (
-    <>
-    {currentView === 'empty' ? (
+    <Stack className="editor" data-testid="item-editor-inner">
+      <HelloWorldItemRibbon
+        {...props}
+        isSaveButtonEnabled={isSaveEnabled()}
+        currentView={currentView}
+        saveItemCallback={SaveItem}
+        openSettingsCallback={handleOpenSettings}
+        navigateToGettingStartedCallback={navigateToGettingStarted}
+      />
+      {currentView === VIEW_TYPES.EMPTY ? (
         <HelloWorldItemEditorEmpty
           workloadClient={workloadClient}
           item={item}
           onNavigateToGettingStarted={navigateToGettingStarted}
         />
       ) : (
-        <HelloWorldItemEditorGettingStarted
+        <HelloWorldItemEditorDefault
           workloadClient={workloadClient}
           item={item}
-          onNavigateToEmpty={navigateToEmpty}
         />
       )}
-    </>
+    </Stack>
   );
 }
