+++ Workload/app/items/HelloWorldItem/HelloWorldItemModel.ts
@@ -1,3 +1,10 @@
 export interface HelloWorldItemDefinition  {
-  message?: string;
+  state?: string;
 }
+
+export const VIEW_TYPES = {
+  EMPTY: 'empty',
+  GETTING_STARTED: 'getting-started'
+} as const;
+
+export type CurrentView = typeof VIEW_TYPES[keyof typeof VIEW_TYPES];
